Centralize test file configuration in support module
